Committing svn revision 32524: get a database pointer in a way that almost forces you to check the pointer fixes ZCP-9575

+++ provider/libserver/ECShortTermEntryIDManager.cpp
@@ -220,7 +220,7 @@ ECRESULT ECShortTermEntryIDManager::IsShortTermEntryId(const entryId *lpEntryId,
 		goto exit;
 	}
 	
-	if (lpEntryId->__size < CbNewEID(""))
+	if (lpEntryId->__size < (int)CbNewEID(""))
 		*lpbResult = false;
 	else {
 		const EID *lpEid = (const EID*)lpEntryId->__ptr;
